Rewrites/improvements in rendering and device handling

+++ RGB.NET.Devices.Corsair/Helper/NativeExtensions.cs
@@ -8,10 +8,10 @@ namespace RGB.NET.Devices.Corsair
         internal static Rectangle ToRectangle(this _CorsairLedPosition position)
         {
             //HACK DarthAffe 08.07.2018: It seems like corsair introduced a bug here - it's always 0.
-            double width = position.width < 0.5 ? 10 : position.width;
-            double height = position.height < 0.5 ? 10 : position.height;
-            double posX = position.left;
-            double posY = position.top;
+            float width = position.width < 0.5f ? 10 : (float)position.width;
+            float height = position.height < 0.5f ? 10 : (float)position.height;
+            float posX = (float)position.left;
+            float posY = (float)position.top;
 
             return new Rectangle(posX, posY, width, height);
         }
